requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django<=4.2.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
